キーボードのイベントをファイル分離
